bug fix row 2

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: blanglai <blanglai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/16 20:27:35 by blanglai          #+#    #+#             */
-/*   Updated: 2025/10/28 09:27:28 by blanglai         ###   ########.fr       */
+/*   Updated: 2025/10/28 09:56:47 by blanglai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
